rename new iti-104 implementation to correct classnames
