--- src/main/java/ch/bfh/ti/i4mi/mag/pmir/iti93/Iti93Constants.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ch.bfh.ti.i4mi.mag.pmir.iti93;
-
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-
-/**
- * @author Oliver Egger
- */
-public interface Iti93Constants {
-
-    Set<String> ITI93_PARAMETERS = new HashSet<>(Arrays.asList(
-            "content",
-            "async",
-            "url"));
-
-    String PMIR_OPERATION_NAME = "$process-message";
-}
